+++ ignite/cmd/bubblemodel/chain_serve.go
@@ -7,7 +7,7 @@ import (
 
 	tea "github.com/charmbracelet/bubbletea"
 
-	"github.com/ignite/cli/v29/ignite/pkg/announcements"
+	"github.com/ignite/cli/v29/ignite/internal/announcements"
 	"github.com/ignite/cli/v29/ignite/pkg/cliui/colors"
 	"github.com/ignite/cli/v29/ignite/pkg/cliui/icons"
 	cliuimodel "github.com/ignite/cli/v29/ignite/pkg/cliui/model"
