+++ ignite/internal/analytics/analytics.go
@@ -12,11 +12,11 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/ignite/cli/v29/ignite/config"
+	"github.com/ignite/cli/v29/ignite/internal/sentry"
 	"github.com/ignite/cli/v29/ignite/pkg/cliui"
 	"github.com/ignite/cli/v29/ignite/pkg/errors"
 	"github.com/ignite/cli/v29/ignite/pkg/matomo"
 	"github.com/ignite/cli/v29/ignite/pkg/randstr"
-	"github.com/ignite/cli/v29/ignite/pkg/sentry"
 	"github.com/ignite/cli/v29/ignite/version"
 )
