+++ ignite/pkg/cosmosgen/generate_typescript.go
@@ -2,19 +2,16 @@ package cosmosgen
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
 	"log"
-	"net/http"
 	"os"
 	"os/exec"
 	"path/filepath"
 	"sort"
 	"strings"
-	"time"
 
 	"golang.org/x/sync/errgroup"
 
+	"github.com/ignite/cli/v29/ignite/internal/buf"
 	"github.com/ignite/cli/v29/ignite/pkg/cache"
 	"github.com/ignite/cli/v29/ignite/pkg/cosmosanalysis/module"
 	"github.com/ignite/cli/v29/ignite/pkg/cosmosbuf"
@@ -30,8 +27,6 @@ var (
 	protocGenTSProtoBin = "protoc-gen-ts_proto"
 
 	msgBufAuth = "Note: Buf is limits remote plugin requests from unauthenticated users on 'buf.build'. Intensively using this function will get you rate limited. Authenticate with 'buf registry login' to avoid this (https://buf.build/docs/generate/auth-required)."
-
-	bufTokenEndpoint = "https://api.ignite.com/v1/buf/token"
 )
 
 const localTSProtoTmpl = `version: v1
@@ -39,9 +34,12 @@ plugins:
   - plugin: ts_proto
     out: .
     opt:
-      - "esModuleInterop=true"
-      - "forceLong=long"
-      - "useOptionals=true"
+    	- logtostderr=true
+    	- allow_merge=true
+    	- json_names_for_fields=false
+    	- ts_proto_opt=snakeToCamel=true
+    	- ts_proto_opt=esModuleInterop=true
+    	- ts_proto_out=.
 `
 
 type tsGenerator struct {
@@ -63,8 +61,7 @@ func newTSGenerator(g *generator) *tsGenerator {
 
 	if !tsg.isLocalProto {
 		if os.Getenv(bufTokenEnvName) == "" {
-			// fetch ignite buf.build token
-			token, err := fetchBufToken()
+			token, err := buf.FetchToken()
 			if err != nil {
 				log.Printf("No '%s' binary found in PATH, using remote buf plugin for Typescript generation. %s\n", protocGenTSProtoBin, msgBufAuth)
 			} else {
@@ -265,32 +262,4 @@ func (g *tsGenerator) generateRootTemplates(p generatePayload) error {
 	}
 
 	return templateTSClientRoot.Write(outDir, "", p)
-}
-
-// fetchBufToken fetches the buf token from the ignite API
-func fetchBufToken() (string, error) {
-	client := &http.Client{
-		Timeout: 10 * time.Second,
-	}
-
-	resp, err := client.Get(bufTokenEndpoint)
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return "", fmt.Errorf("HTTP request failed with status code: %d", resp.StatusCode)
-	}
-
-	type tokenResponse struct {
-		Token string `json:"token"`
-	}
-	var tokenResp tokenResponse
-
-	if err := json.NewDecoder(resp.Body).Decode(&tokenResp); err != nil {
-		return "", err
-	}
-
-	return tokenResp.Token, nil
 }
