+++ ignite/templates/app/files/{{protoDir}}/buf.gen.swagger.yaml
@@ -1,4 +1,4 @@
-# This file is auto-generated from Ignite. You can edit
+# This file is auto-generated by Ignite. You can edit
 # the file content but do not change the file name or path.
 #
 # buf.gen.swagger.yaml
